api/requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-API==2.0
Flask-HTTPAuth==4.1.0
Flask-SQLAlchemy==2.4.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
psycopg2==2.8.5
pydocstyle==5.0.2
PyJWT==1.7.1
pylint==2.5.3
six==1.15.0
snowballstemmer==2.0.0
sqlalchemy==1.3.13
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
